requirements.txt.txt
python-telegram-bot==20.3
openai==1.35.3
requests
